Fixed msvc compilation problems related to glyph indices.
git-svn-id: svn://jarda-home/trunk/jagbase@2234 e431c221-245b-724f-9001-839d8ccdb2d3

+++ code/include/resources/interfaces/typeface.h
@@ -237,7 +237,7 @@ public:
      * @return corresponding gid or 0 if no gid is found (standard convention)
      * Implementations are allowed not to implement this method.
      */
-    virtual Int codepoint_to_gid(Int codepoint) const PURE_FUNCTION;
+    virtual UInt16 codepoint_to_gid(Int codepoint) const PURE_FUNCTION;
     virtual FaceCharIterator char_iterator() const PURE_FUNCTION;
